+++ releasenotes.md
@@ -2,7 +2,8 @@
 
 ## Version 0.18.0
 ### New Features
-- Internet.Email()
+- Internet.Email(string = null)
+- Internet.UserName(string = null)
 
 ## Version 0.17.0
 ### New Features